fix: CMakeLists.txt MSVC-kompatible Compiler-Flags und C4244-Warnungen beheben
- CMakeLists.txt: GCC/Clang-Flags (-g, -O0, -O3) durch if(MSVC)-Block ersetzt;
  MSVC nutzt nun /bigobj /W4 über add_compile_options(), Debug/Release-Optimierung
  wird von CMake automatisch als /Od bzw. /O2 gesetzt (keine D9002-Warnungen mehr)
- CMakeLists.txt: src/FEC/word_interleaver.cpp zur ale_fec-Library ergänzt
- symbol_decoder.cpp: C4244 in bin_to_symbol() und decode_word_with_voting()
  durch explizite static_cast<uint8_t>-Casts behoben
- tone_generator.cpp, ale_waveform.cpp: C4244 (double→float) durch
  static_cast<float> bei sin/cos-Twiddle-Initialisierung behoben

+++ src/FSK/ale_waveform.cpp
@@ -22,8 +22,8 @@ FFTBuffer::FFTBuffer()
 
     for (uint32_t k = 0; k < FFT_SIZE; ++k) {
         double angle = 2.0 * M_PI * k / FFT_SIZE;
-        fft_cs_twiddle[k] = std::cos(angle);
-        fft_ss_twiddle[k] = std::sin(angle);
+        fft_cs_twiddle[k] = static_cast<float>(std::cos(angle));
+        fft_ss_twiddle[k] = static_cast<float>(std::sin(angle));
     }
 }
